messaging_app/requirements.txt
# Core Django and DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-extensions==3.2.3
django-environ==0.11.2

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
cryptography==41.0.7

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Caching & Sessions
redis==5.0.1
django-redis==5.4.0

# Real-time functionality
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# File handling & Storage
Pillow==10.1.0
django-storages==1.14.2

# API Documentation
drf-spectacular==0.27.0
django-rest-swagger==2.2.0

# Search functionality
django-elasticsearch-dsl==8.0.0
elasticsearch==8.11.0

# Email handling
django-ses==3.5.0
sendgrid==6.10.0

# Development & Testing
django-debug-toolbar==4.2.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
faker==20.1.0

# Monitoring & Logging
django-prometheus==2.3.1
django-structlog==5.1.0

# API Documentation & Validation
coreapi==2.3.3
pyyaml==6.0.1

# Additional utilities
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0